refactor(store-core): extract permission-request handlers into handlers/permission.ts (audit P2-3)
Move the permission-request family (handlePermissionRequest, handlePermissionResolved,
handlePermissionExpired, handlePermissionTimeout, handlePermissionRulesUpdated +
the PermissionRule type and their payload interfaces) out of the handlers/index.ts
god-module into handlers/permission.ts, behind the barrel's export * re-export so
the public surface is unchanged.

The permission-MODE handlers (permission_mode_changed / available_permission_modes
/ confirm_permission_mode) stay in index.ts for now — handleClaudeReady consumes
handleAvailablePermissionModes + PermissionMode, so they move in a later slice
once claude_ready is addressed.

Drop the now-unused ChatMessage (../types) and nextMessageId (../utils) imports
from index.ts — both were only used by the moved permission-expired/timeout
auto-deny system messages.

Pure move — body byte-identical to origin/main. index.ts 1334 -> 1165 lines.
